Don't install to $HOME/blah
.. as it's not a native path and the install prefix
must be understood as a native path for Windows
host/build to succeed.

Otherwise, bash is invoked to run the compilers or
assemblers and it seems that MSYS2->Native->MSYS2
can cause crt heap corruption. This issue could be
related to the problems I was having making Linux
targets on Windows? More testing of this theory is
needed."